Kernel v.5.18+ ASHMEM fix
K.ver.15.8+ has dropped ASHMEM completly which makes Anbox unusable. This quick commit fixes it for now for the "user" part of anbox. MEMFD is the successor to ASHMEM which is not supported by Anbox yet - it is until now unclear if it ever will... This fix may break any time with further updates...